requirements.txt
tensorflow
flask
numpy
opencv-python
pillow
pandas
certifi
